// sanity.types.ts
export type Product = {
  _id: string;
  _type: "product";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  price: number;
  description?: string;
  categories?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "category";
  }>;
  versions?: Array<{
    size?: "Small" | "Medium" | "Large";
    color?:
      | "Blue"
      | "Black"
      | "Brown"
      | "Green"
      | "Gray"
      | "Red"
      | "White"
      | "Yellow";
    stock?: number;
    _type: "version";
    _key: string;
  }>;
  images?: Array<{
    color?:
      | "Blue"
      | "Black"
      | "Brown"
      | "Gray"
      | "Green"
      | "Red"
      | "White"
      | "Yellow";
    images?: Array<{
      asset?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
      };
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: "image";
      _key: string;
    }>;
    _key: string;
  }>;
  isTrending?: boolean;
  publishedAt?: string;
  slug?: Slug;
};

export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;